requirements.txt
backcall==0.1.0
bleach==3.0.2
certifi==2018.8.24
chardet==3.0.4
colorama==0.4.0
decorator==4.3.0
defusedxml==0.5.0
Django==2.1.3
django-autocomplete-light==3.3.2
django-crispy-forms==1.7.2
django-extensions==2.1.3
django-filter==2.0.0
django-reversion==3.0.0
django-tables2==2.0.1
djangorestframework==3.8.2
docutils==0.14
entrypoints==0.2.3
future==0.16.0
idna==2.7
ipykernel==5.1.0
ipython==7.1.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
isodate==0.6.0
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
lxml==4.2.5
MarkupSafe==1.1.0
mistune==0.8.4
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.2
numpy==1.15.2
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pickleshare==0.7.5
pkginfo==1.4.2
prometheus-client==0.4.2
prompt-toolkit==2.0.7
Pygments==2.2.0
pyparsing==2.2.2
python-dateutil==2.7.3
pytz==2018.5
pywinpty==0.5.4
pyzmq==17.1.2
qtconsole==4.4.3
rdflib==4.2.2
readme-renderer==23.0
requests==2.20.1
requests-toolbelt==0.8.0
Send2Trash==1.5.0
six==1.11.0
SPARQLWrapper==1.8.2
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
tqdm==4.28.1
traitlets==4.3.2
twine==1.12.1
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
wincertstore==0.2
